Move MacAddrField
